+++ src/components/home/prompt/PromptListSection.tsx
@@ -25,12 +25,11 @@ import { usePathname } from "next/navigation";
 import { memo, useEffect, useRef, useState } from "react";
 import ScrollButton from "@/components/common/ScrollButton/ScrollButton";
 import { boolean } from "zod";
+import useScrollButtonControl from "@/hooks/ui/useScrollButtonControl";
 
 interface PromptListSectionProps {
     viewType?: ViewType;
 }
-
-export type CurrentScrollType = "right" | "left" | "switching";
 
 const PromptListSection = ({ viewType = "open" }: PromptListSectionProps) => {
     const searchedKeyword = useRecoilValue(searchedKeywordState);
@@ -40,84 +39,8 @@ const PromptListSection = ({ viewType = "open" }: PromptListSectionProps) => {
     const limit = isUnderTablet ? 5 : 18;
     const pathname = usePathname();
 
-    const [currentScroll, setCurrentScroll] =
-        useState<CurrentScrollType>("left");
-    const scrollLeftRef = useRef<HTMLDivElement>(null);
-    const scrollRightRef = useRef<HTMLDivElement>(null);
-
-    const handleScroll = () => {
-        if (currentScroll === "left" && scrollRightRef.current) {
-            scrollRightRef.current.scrollIntoView({
-                behavior: "smooth",
-                block: "nearest",
-                inline: "start",
-            });
-            setCurrentScroll("right");
-        }
-        if (currentScroll === "right" && scrollLeftRef.current) {
-            scrollLeftRef.current.scrollIntoView({
-                behavior: "smooth",
-                block: "nearest",
-                inline: "start",
-            });
-            setCurrentScroll("left");
-        }
-    };
-
-    useEffect(() => {
-        const observerOptions = {
-            root: null,
-            rootMargin: "0px",
-            threshold: 0.7, // 요소가 90% 이상 보일 때 감지
-        };
-        console.log("감지중");
-
-        const handleIntersection = () => {
-            const leftVisible =
-                scrollLeftRef.current?.getBoundingClientRect().left ?? -1 >= 0;
-            const rightVisible =
-                scrollRightRef.current?.getBoundingClientRect().right ??
-                window.innerWidth <= window.innerWidth;
-
-            if (leftVisible && !rightVisible) {
-                setCurrentScroll("left");
-                console.log("set to left");
-            } else if (!leftVisible && rightVisible) {
-                setCurrentScroll("right");
-                console.log("set to right");
-            } else {
-                setCurrentScroll("switching");
-                console.log("set to switching");
-            }
-        };
-
-        const leftObserver = new IntersectionObserver(([entry]) => {
-            if (entry.isIntersecting) {
-                setCurrentScroll("left");
-                console.log("90% left, set to left");
-            } else {
-                handleIntersection();
-            }
-        }, observerOptions);
-
-        const rightObserver = new IntersectionObserver(([entry]) => {
-            if (entry.isIntersecting) {
-                setCurrentScroll("right");
-                console.log("90% right, set to right");
-            } else {
-                handleIntersection();
-            }
-        }, observerOptions);
-
-        if (scrollLeftRef.current) leftObserver.observe(scrollLeftRef.current);
-        if (scrollRightRef.current)
-            rightObserver.observe(scrollRightRef.current);
-
-        return () => {
-            leftObserver.disconnect();
-            rightObserver.disconnect();
-        };
-    }, []);
+    const { scrollLeftRef, scrollRightRef, handleScroll, currentScroll } =
+        useScrollButtonControl();
 
     const promptContent = () => {
         if (searchedKeyword && pathname === "/") {
